# Implement fast pass for CPU scalars /number literals